Creating initial "exec" and "call" functions
These are like Scheme's "eval" and "apply", which serve as the base interpreter for the whole system. These functions are directly visible to interpreted code, which can invoke them or inspect/modify them as objects within the root object (window.Objects, which is an object representing the whole of the system). Ultimately, the these (and other) functions will be replaced (via bootstrap code) with code made up of the very objects that the system interprets, and thus the system will become self-defining and capable of self-modification AT runtime.