Updating Objects.html with TODO to match branch
Right now, the specific branch is hard-coded into the URL to dynamically load the POC code (Objects.js) into this HTML previewer. I should update it to programmatically determine the URL for Objects.js based on the current URL from which Objects.html was loaded. Then I will not need to modify this URL every time I branch or merge.