Created EvalGrammar.txt as a help to make eval()
After tracing the current eval (O.eval) code, I've found inconsistencies and other holes. So, I made this "grammar" to show what the result should be for each class of thing being eval'd, partly as documentation, but mainly as a guide to make sure I code it properly. In the process though, I've reformed these rules, and I need to update eval to work according to these new rules. For example, {op:x, args:[a..]} is replaced with ["x", a..], and I've decided to distinguish between code expressions [op, arg1, arg2..] and func declarations {args:["name1", "name2"..], body:[..]}, so that there is a way to determine how to map which arguments to what names, and when to create a new "child" env (context) from the parent env.

I intend to re-do this, as the entire set of rules has to be extrapolated from these all-in-one rules. I will split it into several sub-functions for evaluation to help cover the full spec, though I might actually still CODE it as one single function.